refactor: 불필요한 ModuleInjector 삭제
